+++ modules/ec2/ec2-auto-scaling-group/_ec2_var.tf
@@ -45,8 +45,13 @@ variable "instance_type" {
   type        = string
 }
 
-variable "volume_size" {
-  description = "The size of the EBS volume (in GB)"
+variable "root_volume_size" {
+  description = "The size of the EBS volume (in GB) for root partition"
+  type        = number
+}
+
+variable "data_volume_size" {
+  description = "The size of the EBS volume (in GB) for additional partition"
   type        = number
 }