+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Production Grade Modularity AWS Terraform Modules
 
 ![all contributors](https://img.shields.io/github/contributors/devopscorner/tf-aws-modules)
 ![tags](https://img.shields.io/github/v/tag/devopscorner/tf-aws-modules?sort=semver)
+![terraform](https://img.shields.io/static/v1?logo=terraform&logoColor=white&label=terraform&message=1.5.2&color=orange)
 ![download all](https://img.shields.io/github/downloads/devopscorner/tf-aws-modules/total.svg)
 ![view](https://views.whatilearened.today/views/github/devopscorner/tf-aws-modules.svg)
 ![clone](https://img.shields.io/badge/dynamic/json?color=success&label=clone&query=count&url=https://raw.githubusercontent.com/devopscorner/tf-aws-modules/master/clone.json?raw=True&logo=github)
@@ -14,6 +15,17 @@ Production Grade Modularity AWS Terraform Modules
 [![license](https://img.shields.io/github/license/devopscorner/tf-aws-modules)](https://img.shields.io/github/license/devopscorner/tf-aws-modules)
 
 ---
+
+## Version 1.1.0
+
+- Terraform Workspace Example
+  - Terraform Workspace Module EC2 Amazon Linux 2
+  - Terraform Workspace Module EC2 Amazon Linux 2023
+  - Terraform Workspace Module EC2 Ubuntu
+  - Terraform Workspace Module EC2 Autoscaling (Ubuntu 22.04)
+  - Terraform Workspace Module EC2 for Airflow
+  - Terraform Workspace Module EC2 for Jumphost
+
 ## Version 1.0.0
 
 ### Features